requirements/base.txt
Django==2.1.1
djangorestframework==3.8.2

django-filter==2.0.0

djangorestframework-jwt==1.11.0
PyJWT==1.6.4

environ==1.0
Pillow==5.2.0
pytz==2018.5

# PostgreSQL
psycopg2==2.7.5
psycopg2-binary==2.7.5
